+++ Root/RegionVarCalculator_tls.cxx
@@ -26,7 +26,7 @@ EL::StatusCode RegionVarCalculator_tls::doInitialize(EL::Worker * worker) {
 }
 
 EL::StatusCode RegionVarCalculator_tls::doCalculate(std::map<std::string, double              >& RegionVars,
-						     std::map<std::string, std::vector<double> >& VecRegionVars){
+						     std::map<std::string, std::vector<float> >& VecRegionVars){
   xAOD::TStore * store = m_worker->xaodStore();//grab the store from the worker
   xAOD::TEvent* event = m_worker->xaodEvent();
 
@@ -52,7 +52,7 @@ EL::StatusCode RegionVarCalculator_tls::doCalculate(std::map<std::string, double
 
 
 EL::StatusCode RegionVarCalculator_tls::doAllCalculations(std::map<std::string, double>& RegionVars,
-							   std::map<std::string, std::vector<double> > & VecRegionVars)
+							   std::map<std::string, std::vector<float> > & VecRegionVars)
 {
   xAOD::TStore * store = m_worker->xaodStore();
   xAOD::TEvent * event = m_worker->xaodEvent();
@@ -76,10 +76,10 @@ EL::StatusCode RegionVarCalculator_tls::doAllCalculations(std::map<std::string, 
   STRONG_CHECK(store->retrieve(jets_nominal, "selectedJets"));
 
   //  const std::vector<xAOD::IParticle*> & jetStdVec = jetcont->stdcont();
-  std::vector<double> jetPtVec;
-  std::vector<double> jetEtaVec;
-  std::vector<double> jetPhiVec;
-  std::vector<double> jetEVec;
+  std::vector<float> jetPtVec;
+  std::vector<float> jetEtaVec;
+  std::vector<float> jetPhiVec;
+  std::vector<float> jetEVec;
 
   for( const auto& jet : *jets_nominal) {
     jetPtVec.push_back( toGeV(jet->pt()));
@@ -96,11 +96,11 @@ EL::StatusCode RegionVarCalculator_tls::doAllCalculations(std::map<std::string, 
   xAOD::IParticleContainer* leptons_nominal(nullptr);
   STRONG_CHECK(store->retrieve(leptons_nominal, "selectedLeptons"));
 
-  std::vector<double> lepPtVec;
-  std::vector<double> lepEtaVec;
-  std::vector<double> lepPhiVec;
-  std::vector<double> lepEVec;
-  std::vector<double> lepPdgidVec;
+  std::vector<float> lepPtVec;
+  std::vector<float> lepEtaVec;
+  std::vector<float> lepPhiVec;
+  std::vector<float> lepEVec;
+  std::vector<float> lepPdgidVec;
 
   auto lepPdgId = [](xAOD::IParticle* p){
     xAOD::Electron * el = dynamic_cast<xAOD::Electron*>(p);
@@ -134,7 +134,7 @@ EL::StatusCode RegionVarCalculator_tls::doAllCalculations(std::map<std::string, 
 
 
 EL::StatusCode RegionVarCalculator_tls::doSRCalculations(std::map<std::string, double>& RegionVars,
-							  std::map<std::string, std::vector<double> > & VecRegionVars)
+							  std::map<std::string, std::vector<float> > & VecRegionVars)
 {
 
   xAOD::TStore * store = m_worker->xaodStore();
@@ -169,10 +169,10 @@ EL::StatusCode RegionVarCalculator_tls::doSRCalculations(std::map<std::string, d
 
 
 EL::StatusCode RegionVarCalculator_tls::doCR1LCalculations(std::map<std::string, double>& RegionVars,
-							    std::map<std::string, std::vector<double> > & VecRegionVars)
+							    std::map<std::string, std::vector<float> > & VecRegionVars)
 {/*todo*/return EL::StatusCode::SUCCESS;}
 
 
 EL::StatusCode RegionVarCalculator_tls::doCR0LCalculations(std::map<std::string, double>& RegionVars,
-							    std::map<std::string, std::vector<double> > & VecRegionVars)
+							    std::map<std::string, std::vector<float> > & VecRegionVars)
 {/*todo*/return EL::StatusCode::SUCCESS;}
